chore: create node_data/inner_node
