+++ crates/committer/src/patricia_merkle_tree.rs
@@ -1,5 +1,6 @@
 pub mod errors;
 pub mod filled_tree;
+pub mod node_data;
 pub mod original_skeleton_tree;
 pub mod types;
 pub mod updated_skeleton_tree;
