+++ crates/committer/src/patricia_merkle_tree/filled_tree/node_serde.rs
@@ -1,8 +1,10 @@
 use crate::hash::hash_trait::HashOutput;
-use crate::patricia_merkle_tree::filled_tree::node::{BinaryData, FilledNode, LeafData, NodeData};
+use crate::patricia_merkle_tree::filled_tree::node::{FilledNode, LeafData};
 use crate::patricia_merkle_tree::filled_tree::tree::FilledTreeResult;
+use crate::patricia_merkle_tree::node_data::inner_node::{
+    BinaryData, EdgeData, EdgePath, EdgePathLength, NodeData, PathToBottom,
+};
 use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTreeResult;
-use crate::patricia_merkle_tree::types::{EdgeData, EdgePath, EdgePathLength, PathToBottom};
 use crate::storage::storage_trait::{create_db_key, StorageKey, StorageValue};
 use crate::types::Felt;
 use serde::{Deserialize, Serialize};
