+++ crates/committer/src/patricia_merkle_tree/node_data.rs
@@ -1,0 +1,1 @@
+pub mod inner_node;
