+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/node.rs
@@ -1,5 +1,6 @@
 use crate::hash::hash_trait::HashOutput;
-use crate::patricia_merkle_tree::types::{EdgeData, LeafDataTrait, PathToBottom};
+use crate::patricia_merkle_tree::node_data::inner_node::{EdgeData, PathToBottom};
+use crate::patricia_merkle_tree::types::LeafDataTrait;
 
 #[allow(dead_code)]
 #[derive(Debug, PartialEq, Eq)]
