+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/original_skeleton_calc.rs
@@ -1,12 +1,12 @@
 use crate::hash::hash_trait::HashOutput;
-use crate::patricia_merkle_tree::filled_tree::node::BinaryData;
 use crate::patricia_merkle_tree::filled_tree::node::FilledNode;
 use crate::patricia_merkle_tree::filled_tree::node::LeafData;
-use crate::patricia_merkle_tree::filled_tree::node::NodeData;
+use crate::patricia_merkle_tree::node_data::inner_node::BinaryData;
+use crate::patricia_merkle_tree::node_data::inner_node::EdgeData;
+use crate::patricia_merkle_tree::node_data::inner_node::NodeData;
+use crate::patricia_merkle_tree::node_data::inner_node::PathToBottom;
 use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTree;
 use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTreeResult;
-use crate::patricia_merkle_tree::types::EdgeData;
-use crate::patricia_merkle_tree::types::PathToBottom;
 use crate::patricia_merkle_tree::types::TreeHeight;
 use crate::patricia_merkle_tree::updated_skeleton_tree::tree::UpdatedSkeletonTreeImpl;
 use crate::patricia_merkle_tree::{
