+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree/tree.rs
@@ -4,13 +4,13 @@ use std::sync::{Arc, Mutex};
 use crate::hash::hash_trait::{HashFunction, HashOutput};
 use crate::patricia_merkle_tree::errors::UpdatedSkeletonTreeError;
 use crate::patricia_merkle_tree::filled_tree::tree::FilledTree;
+use crate::patricia_merkle_tree::node_data::inner_node::{BinaryData, EdgeData, NodeData};
 use crate::patricia_merkle_tree::types::{LeafDataTrait, NodeIndex, TreeHashFunction};
 use crate::patricia_merkle_tree::updated_skeleton_tree::node::UpdatedSkeletonNode;
 use crate::types::Felt;
 
-use crate::patricia_merkle_tree::filled_tree::node::{BinaryData, FilledNode, NodeData};
+use crate::patricia_merkle_tree::filled_tree::node::FilledNode;
 use crate::patricia_merkle_tree::filled_tree::tree::FilledTreeImpl;
-use crate::patricia_merkle_tree::types::EdgeData;
 
 #[cfg(test)]
 #[path = "tree_test.rs"]
